feat: 启动器主页选项
